Fixed current rate conversion

+++ wildberriesKosmos/getNewOrders.php
@@ -72,7 +72,7 @@ foreach ($newOrders as $newOrder)
 	$position ['quantity'] = 1;
 	$position ['reserve'] = 1;
 	$currentcyRate = $newOrder['convertedPrice'] / $newOrder['price'];
-	$position ['price'] = (isset($newOrder['salePrice']) && $newOrder['salePrice'] != null ? $newOrder['salePrice'] : $newOrder['price'] * $currentcyRate);
+	$position ['price'] = (int)((isset($newOrder['salePrice']) && $newOrder['salePrice'] != null ? $newOrder['salePrice'] : $newOrder['price']) * $currentcyRate);
 	$position ['assortment'] = array(
 		'meta' => $productMS[0]['meta']
 	);
